Build Example Registry Token Server
This commit builds the example token server from the Docker
Distribution code base. We will eventually modify this to use
Openstack Keystone as a back end.

For now, the example token server is copied as is. We do not
want to clean up needless code before changing the back end to
Keystone because we might end up ripping out something we need later.
The src also includes the registry, which we will clean up after,
since we are leveraging the existing Swift support for Keystone.

The example token server uses htpasswd as a back end. A htpasswd file
with username and password testuser and testpassword is installed by
default. This will go away after we switch the back end from htpasswd
to Keystone.

Task: 26645
Story: 2002840

Change-Id: I162ff6fdcf61f9bdbc563de52c74f94654557ba7
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>