Uprev kubernetes from 1.10 to 1.12.1
This requires golang upversioned to 1.10.2 or later

The older kubernetes was a src rpm. This is a src tarball
so the files and patches from the old src rpm are included here.

Some of the files from the src rpm failed lint checks, so
they have been updated.

ppc patch from 1.10 was removed.

The kubeadm.conf originates  from:
 kubernetes/build/rpms/10-kubeadm.conf

Story: 2002843
Task: 26751
Change-Id: Ib5602a4a57c0e8998430b0774375853cdd2bca01
Depends-On: I1ea17cc89d839e0ffa90459965c521e504695294
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>