remove lshell
There is security related issue with lshell, and it is not maintained now.
So remove it from our system to avoid security issue.

lshell is used in ldapscripts. A patch is added to switch to bash instead
of lshell. Also the wrs.sudo file in lshell is moved to ldapscripts.

Closes-Bug: 1795451

Change-Id: Ia5de1bc94d22eb6c9bea6d9a96e92564ad848b19
Signed-off-by: slin14 <shuicheng.lin@intel.com>