refactor shadow-utils
The refactor task is to eliminate patch which is relate to configure
file modification. And we could replace srpm with rpm if all patches
could be eliminated, which will help save build time.
To eliminate the patch, a new config package or puppet file will be
created to do the configure file modification. Here is the general
guide for the refactor process:
https://review.openstack.org/612292

For shadow-utils, the configure file is moved to shadow-utils-config
package, so src rpm is replaced with rpm.
shadow-utils-config is configured to depend on shadow-utils, so
shadow-utils rpm will be picked automatically in ISO build.
File clear_shadow_locks and su under files folder is not used in build,
so just delete them.

This patch should has no function change from end user view.
Test has been done:
Done build and deploy test, and confirm login.defs and
clear_shadow_locks.service in the deploy node are the same as before.

Story: 2003768
Task: 27593

Change-Id: Ie62800c64aca54c39950266f8cf36e47cc1f55bd
Signed-off-by: slin14 <shuicheng.lin@intel.com>