Port: Ceph recovery logic improvements for blocked ops
Our current version of Jewel LTS (v10.2.6) has an issue on cluster
recovery where we will see blocked operations associated with an OSD
under certain conditions.

The cluster will remain in a HEALTH_WARN state and exhibit the following
warning:
       1 ops are blocked > 1048.58 sec on osd.1

This will block patching orchistration. This commit enhances the
recovery code so that we will look for this condition and restart the
OSD that sees these blocked operations.

Based on only having 1 peer group with co-located HDD based journals,
recovery can take longer than when we have SSD journals in play.
Increase the threshold that triggers a restart to provide additional
time for the OSDs to start, execute recovery, and allow the cluster to
settle into a steady state. This is also well within the orchestration
window before the wait-data-sync timeout occurs

use systemd-coredump instead of "gcore" to avoid manual
disk space management of core files

Story: 2003605
Task: 28857

Change-Id: Ib0dc595b2ff24fcbd3bbe1dd8215a7a183a5554c
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>
Signed-off-by: Changcheng Liu <changcheng.liu@intel.com>