refactor lighttpd
Package lighttpd-config is added to config customized config
file of lighttpd.

lighttpd-csr.conf is removed, since it is not used in build.
Patches in files folder are moved to patches folder.

Test:
Pass build and multi node deploy test. Confirmed related config
file is the same as before in deploy.

Story: 2003768
Task: 27616
Depends-On: https://review.openstack.org/617221

Change-Id: I9155542754985429f06fa569f3644009a39ab8c1
Signed-off-by: slin14 <shuicheng.lin@intel.com>