Remove hardcoded configure options for openldap
The openldap-spec-file.patch contains some modifications to the
default configure command line.
After evaluated by Saul in task 27731, we should be able to remove
related changes in this patch.
It seems some changes could not be removed, so the patch could not
be reverted so far.

Deployment test pass and slapd service works.

Story: 2004216
Task: 28015

Signed-off-by: zhipengl <zhipengs.liu@intel.com>

Change-Id: I55e4961bf2ceb69bb0592f3fb34b4fded3a2e8fd