Traffic control: fix TC filters for vlan sub-interface
Sometime after kernel 3.10.0-514.16.1.X, tc filter commands no longer
match 802.1q packets when the filter protocol is set to 'ip'.

This poses a problem for a consolidated (eg. infra w/ vlan over management)
interface configuration.

The tc filter will operate properly on the vlan interface, but all traffic
will go to the default qdisc (low priority) when it arrives with a vlan
tag at the sub-interface.

This commit sets the filter protocol to '802.1q' in the case of a subinterface
with a vlan tagged interface ontop of it.

Closes-Bug: #1807055
Change-Id: I457faa2b56bbd270c104cc0313ffe3cc1bfd4db3
Signed-off-by: Steven Webster <steven.webster@windriver.com>