update /etc/pam.d/system-auth conent with system-auth.pam
LDAP sudo user not able to login by password.
root cause is that password rules in system-auth is not updated
correctly because system-auth.apm in pam-config is missed to be
copied to /etc/pam.d/system-auth

copy system-auth.pam in pam-config to /etc/pam.d/system-auth to
solve this issue.

Closes-Bug: #1806977

Change-Id: Ic646e30d06bcbe8cf3bf66c903942e4240bd23bd
Signed-off-by: Sun Austin <austin.sun@intel.com>