Port: Update ceph initialization code to work with SM and 2 node configuration
This commit adapts ceph_pmon_wrapper script to work with SM. It adds
possibility to use environment variables as arguments to the script
and renames the script so that it can be installed in /etc/init.d.

SM can pass only a single CLI arguments to init scripts (usually
'start', 'stop' and 'status') yet it can pass any environment variables
to the script. In this commit we update ceph's init wrapper to take
and environment variable "ARGS" and append it to the list of
arguments. E.g. passing ARGS=mon when the script is called
with '/etc/init.d/ceph-init-wrapper start' is equivalent to
/etc/init.d/ceph-init-wrapper start mon'.

No changes to the upstream Ceph code are done.

Story: 2003605
Task: 28857

Change-Id: I418c0e07c506b4e8f4164d08595b53654de34c1c
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>