Add go "dep" tool to support lockfiles in go pkgs
The go compiler supports "go dep" to download dependancies, but
it will download the latest version of dependancies from the git
repository of the dependancy.  This could lead to non-reproducable
builds, or the potential of dependancy API changes (including
deprecated features) to break the build of packages.

The "dep" tool allows for a lockfile to be created, which captures
the git commit ID of dependancies, and can restore dependancies to
the captured state.

Since no binary package containing "dep" is currently availabe for
CentOS, this change adds the package and compiles it from source.

Depends-On: https://review.openstack.org/625333

Story: 2004617
Task: 28525

Change-Id: Ic3bfad6a419b344e7c147c3839d16956c432a99e
Signed-off-by: Jason McKenna <jason.mckenna@windriver.com>