Resolve AIO-SX shutdown hang with CEPH ordering hooks
In ceph-10.2.6, the ceph init script uses systemd-run to launch
ceph-mon and ceph-osd services. This generates transient systemd
service files with basic configuration. On node shutdown, ceph is
getting shutdown while it is still in use by containers, and without
unmapping the RBD devices, causing the libceph kernel module to
hang trying to communicate with the ceph monitor.

This update patches the ceph init script to generate systemd
overrides config files for the ceph-mon and ceph-osd that provide
improved ordering during shutdown, as well as a script that runs
as part of ceph-mon shutdown (ahead of ceph-osd) to unmap the RBD
devices. This ordering ensures kubelet and docker services are
shutdown first, then the ceph-mon and ceph-osd services, followed
by the shutdown of service management (SM). Once kubelet and docker
have shut down, the ceph-preshutdown.sh script is able to cleanly
unmount and unmap the RBD devices and unload the rbd and libceph
kernel modules.

In ceph-11.0.1, the use of systemd-run was replaced with proper
systemd service configuration files. Once ceph is upgraded for
StarlingX, the ordering and cleanup will need to be revisited.

Story: 2004520
Task: 28258
Change-Id: I6f7d7b9e704121c54211afd86b38df015b8d7a63
Signed-off-by: Don Penney <don.penney@windriver.com>