Fix racing caused by /etc/init.d/ceph-init-wrapper operation overlaps
This commit fixes a racing issue between SM's PID monitoring and
'worker' puppet manifests on AIO-DX triggering a restart of ceph-mon.

After a reboot on AIO DX, SM takes control of ceph-mon monitoring
after 'controller' puppet manifests finish applying. As part of this,
SM monitors processes death notification and gets the pid from the
.pid file. And periodically executes '/etc/init.d/ceph status
mon.controller' for a more advanced monitoring.

After the 'controller' puppet manifest finish applying a 2nd set is
executed: the 'worker' puppet manifests. This set triggers a restart
of ceph-mon through /etc/init.d/ceph restart that has two steps: 'stop'
in which ceph-mon is stopped, and 'start' in which it is restarted.

In the first step, stopping ceph-mon leads to the removal of the PID
file. This is promptly detected by SM which immediately triggers a
start of ceph-mon that creates a new pid file. Problem is that
ceph-mon was already in a restart, and at the end of the 'stop' step
the init script cleans up the new pid file instead of the old.

This leads to controllers swacting a couple of times before the system
gets rid of the rogue process.

The commit fixes the race issue by avoiding overlaps between start,
stop & restart operations in /etc/init.d/ceph-init-wrapper.

Change-Id: I2f9ce8d6beeb54358363fc861654ef94c629f6b6
Story: 2002844
Task: 29214
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>