upgrade openvswitch from 2.9.0 to 2.11.0 version
DPDK is upgraded from 17.11 to 18.11 together with openvswitch.
With this upgrade, DPDK for Mellanox NIC support is enabled again,
which is disabled during CentOS 7.6 upgrade, due to DPDK 17.11 cannot
pass build with 7.6 kernel.
Due to srpm for openvswitch 2.11 is not available in centos repo yet,
tarball is used instead. The spec file and other configuration file
from 2.9.0 srpm are kept and updated for 2.11.0 openvswitch.
Meta patches are applied to the spec file directly.
For the 3 patches, 2 of them are abandoned since new DPDK already
contain them. Another one is rebased for 2.11 version.

Test:
Pass build and basic deploy test. ovs-vsctl shows vswitch status is
correct.

Depends-On: https://review.openstack.org/640244
Story: 2004743
Task: 29196

Change-Id: I1fc28f1c7f93e1da2f8e7ac0dfb5d8e294d53c90
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>