Update for systemd patches (CVE-2018-15688)
This commit is a complement for the change in the stx-tools repo
(link: https://review.openstack.org/#/c/651340/), the upgrade is
required to fix the CVE-2018-15688.

The fix was done in Jan 2019 by the community and requires the
update of systemd packages.

System-config needs to be updated in order to be aligned with the
new version of systemd-219-62.el7_6.5.

systemd-config.spec:
	Update 'BuildRequires' value for the new systemd version.

update-package-versioning-for-TIS-format.patch:
	Update 'Release' value for the new version of systemd.

0703-fix-build-error-for-CentOS7.6.patch:
	Update process_http_upload function in order to support the
	new type of value '-ENOBUFS' and changes in the index file.

srpm_path:
	Update the 'mirror' value with the new systemd version.

Change-Id: I8f69532833c216f7a369f1b0596dc6607b0f2a7c
Closes-Bug: #1820756
Depends-On: 651340
Signed-off-by: Mawrer Ramirez <mawrer.a.ramirez.martinez@intel.com>