replace python-gunicorn srpm with rpm
For the source patch 0001-add-worker-abort-hook.patch, it is not used.
For the 3 meta patches, only 0001-TIS-gunicorn-19-upgrade.patch is
applied to srpm. And this meta patch just add tis version, and disable
python3 compilation. So it is ok to drop all these patches, and use
rpm to replace srpm.

Story: 2005484
Task: 30578
Depends-On: https://review.openstack.org/653908

Change-Id: I5805cce1dc64dd41be9f415bab985b34791f1506
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>